refactor: extract NodeInfoBody and unify node detail display
Share a single NodeInfoBody component between the Nodes panel modal
and the Map marker popup so both surfaces always render the same data.
Map popup is styled with a dark theme to match the app.

Fixes #37
